fix: incorrect type for v fixes https://github.com/rarible/protocol-issues/issues/107

+++ nft/core/src/main/kotlin/com/rarible/protocol/nft/core/converters/dto/TokenIdDtoConverter.kt
@@ -13,7 +13,7 @@ object TokenIdDtoConverter : Converter<SignedTokenId, NftTokenIdDto> {
         return NftTokenIdDto(
             tokenId = source.tokenId.value,
             signature = NftSignatureDto(
-                v = source.sign.v,
+                v = source.sign.v.toInt(),
                 r = Binary(source.sign.r),
                 s = Binary(source.sign.s)
             )
